@@ -8,6 +8,7 @@
 	"type": "module",
 	"main": "index.js",
 	"license": "MIT",
+	"packageManager": "bun@1.1.26",
 	"scripts": {
 		"tdd": "vitest",
 		"test": "vitest run",
@@ -20,32 +21,34 @@
 		"tailwind-merge": "2.5.2"
 	},
 	"devDependencies": {
+		"@eslint/eslintrc": "3.1.0",
+		"@eslint/js": "9.9.1",
+		"@kavsingh/eslint-plugin-filenames": "2.0.0-rc.0",
 		"@tailwindcss/container-queries": "0.1.1",
 		"@testing-library/dom": "10.4.0",
 		"@testing-library/jest-dom": "6.5.0",
 		"@testing-library/user-event": "14.5.2",
 		"@tsconfig/strictest": "2.0.5",
 		"@types/bun": "1.1.8",
-		"@types/eslint": "8.56.10",
-		"@typescript-eslint/eslint-plugin": "8.3.0",
-		"@typescript-eslint/parser": "8.3.0",
+		"@types/eslint__eslintrc": "2.1.2",
+		"@types/eslint__js": "8.42.3",
 		"@vitest/coverage-v8": "2.0.5",
 		"autoprefixer": "10.4.20",
-		"eslint": "8.57.0",
+		"eslint": "9.9.1",
 		"eslint-config-prettier": "9.1.0",
 		"eslint-import-resolver-typescript": "3.6.3",
-		"eslint-plugin-deprecation": "3.0.0",
-		"eslint-plugin-filenames": "1.3.2",
 		"eslint-plugin-import-x": "4.1.1",
 		"eslint-plugin-jest-dom": "5.4.0",
 		"eslint-plugin-prettier": "5.2.1",
 		"eslint-plugin-tailwindcss": "3.17.4",
 		"eslint-plugin-testing-library": "6.3.0",
 		"eslint-plugin-vitest": "0.5.4",
+		"globals": "15.9.0",
 		"jsdom": "25.0.0",
 		"prettier": "3.3.3",
 		"tailwindcss": "3.4.10",
 		"typescript": "5.5.4",
+		"typescript-eslint": "8.3.0",
 		"vite": "5.4.2",
 		"vite-plugin-checker": "0.7.2",
 		"vite-tsconfig-paths": "5.0.1",
